requirements.txt
anyio==4.3.0
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
cssselect==1.2.0
Django==5.0.3
django-encrypted-fields==1.1.2
django-jazzmin==2.6.0
django-phonenumber-field==7.3.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jmespath==1.0.1
lxml==5.1.0
packaging==24.0
parsel==1.9.0
phonenumbers==8.13.32
pillow==10.2.0
pyasn1==0.5.1
pycparser==2.21
pycrypto==2.6.1
python-dotenv==1.0.1
python-keyczar==0.716
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
w3lib==2.1.2
